Elaborate on example
I had trouble with the example until I realized that `packer console` works in _JSON mode_ by default and you have to use the option `--config-type=hcl2` so switch to _HCL2 mode_ and enable a function like `timestamp()`.  See [my post to the community](https://discuss.hashicorp.com/t/timestamp-template-function/23341/2).

It might be appropriate to add this to other examples as well.